+++ features/support/preview_checker.rb
@@ -13,9 +13,6 @@ module PreviewChecker
     @topic.split("\n").each do |line|
       expect(page.body).to include(line)
     end
-  end
-
-  def check_purpose
     @purpose.split("\n").each do |line|
       expect(page.body).to include(line)
     end
@@ -39,10 +36,9 @@ module PreviewChecker
 
   def check_storing
     expect(page).to have_tag('p', text: "The data will be kept for #{@shared_duration}.")
-    expect(page).to have_tag('p', text: @shared_usage)
   end
 
-  def check_time_equipment
+  def check_where_when
     expect(page.body).to include('The session is on')
     expect(page).to have_tag('a.editable', text: @session_duration)
     expect(page).to have_tag('a.editable', text: @session_location)
@@ -60,7 +56,7 @@ module PreviewChecker
     )
   end
 
-  def check_incentive
+  def check_incentives
     expect(page.body).to include('We are offering')
     expect(page).to have_tag('a.editable', text: 'a cash incentive of £10.50')
     expect(page.body).to include('for participation in this session')