+++ spec/factories/research_session.rb
@@ -15,16 +15,11 @@ FactoryBot.define do
       step_researcher
       status :topic
       topic 'A topic'
-    end
-
-    trait :step_purpose do
-      step_topic
-      status :purpose
       purpose 'A purpose'
     end
 
     trait :step_methodologies do
-      step_purpose
+      step_topic
       status :methodologies
       methodologies %w[interview survey]
     end
@@ -42,16 +37,16 @@ FactoryBot.define do
       shared_duration '1 year'
     end
 
-    trait :step_time_equipment do
+    trait :step_where_when do
       step_storing
-      status :time_equipment
+      status :where_when
       when_text '1st Sep 2017'
       duration '1 week'
       participant_equipment 'A coat'
     end
 
     trait :step_expenses do
-      step_time_equipment
+      step_where_when
       status :expenses
       travel_expenses_limit '10.00'
       food_expenses_limit '20.00'
@@ -60,16 +55,16 @@ FactoryBot.define do
       food_provided 'Light canapés'
     end
 
-    trait :step_incentive do
+    trait :step_incentives do
       step_expenses
-      status :incentive
+      status :incentives
       incentive true
       payment_type :cash
       incentive_value '40.00'
     end
 
     trait :previewable do
-      step_incentive # really just a readable alias
+      step_incentives # really just a readable alias
     end
 
     trait :nil_expenses do