+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.4.0 / Unreleased
+
+* [FEATURE] - [Add location input to where&when](https://trello.com/c/wdLqH6qT/228-1-add-location-input-to-wherewhen)
+* [FEATURE] - [Create page titles for each step](https://trello.com/c/iZY1S1ZC/200-3-create-page-titles-for-each-step)
+* [FEATURE] - [Merge Topic / Purpose and update copy](https://trello.com/c/YIQw4isH/208-3-merge-topic-purpose-and-update-copy)
+* [FEATURE] - [Navigation improvements](https://trello.com/c/nVXG9lGt/144-navigation-improvements)
+* Remove 'start' step from progress bar
+* Rename steps: 'Data' to 'storing', 'Time & equipment' to 'Where & when', and 'Incentive' to 'Incentives'
+* Progress bar should reflect latest design
+* [FEATURE] - [Share / Save consent form session](https://trello.com/c/zU3zqiy1/92-5-share-save-consent-form-session)
+* [BUG] - [Expenses / refreshments improvements](https://trello.com/c/2mT0SkHT/212-2-expenses-refreshments-improvements)
+
 # 0.3.0 / 2017-10-30
 
 * [BUG] - [Returning to previous steps in the preview bar disables links to subsequent steps](https://trello.com/c/0k7u2liM/199-bug-returning-to-previous-steps-in-the-preview-bar-disables-links-to-subsequent-steps)